+++ Source/Core/Visualization/Renderer/Axis2D.cpp
@@ -1,4 +1,11 @@
+/*****************************************************************************/
+/**
+ *  @file   Axis2D.cpp
+ *  @author Naohisa Sakamoto
+ */
+/*****************************************************************************/
 #include "Axis2D.h"
+#include "ValueAxis.h"
 #include <kvs/OpenGL>
 #include <kvs/Camera>
 #include <kvs/Light>
@@ -7,7 +14,6 @@
 #include <kvs/RGBColor>
 #include <kvs/RGBAColor>
 #include <kvs/TableObject>
-#include "ValueAxis.h"
 
 
 namespace kvs
@@ -23,18 +29,17 @@ Axis2D::Axis2D():
     m_bottom_margin( 30 ),
     m_left_margin( 30 ),
     m_right_margin( 30 ),
+    m_title( "" ),
+    m_title_font(),
+    m_title_offset( 5 ),
     m_background_color( kvs::RGBColor::White() ),
     m_borderline_color( 0, 0, 0, 0.0f ),
     m_borderline_width( 0.0f ),
     m_x_axis( new kvs::ValueAxis( kvs::ValueAxis::Bottom ) ),
-    m_y_axis( new kvs::ValueAxis( kvs::ValueAxis::Left ) ),
-    m_axis_width( 2.0f ),
-    m_axis_color( kvs::RGBColor::Black() ),
-    m_value_color( kvs::RGBColor::Black() ),
-    m_label_color( kvs::RGBColor::Black() ),
-    m_x_label( "" ),
-    m_y_label( "" )
+    m_y_axis( new kvs::ValueAxis( kvs::ValueAxis::Left ) )
 {
+    m_title_font.setStyleToBold();
+    m_title_font.setSize( 22 );
 }
 
 /*===========================================================================*/
@@ -50,7 +55,7 @@ Axis2D::~Axis2D()
 
 /*===========================================================================*/
 /**
- *  @brief  Render 2D axis.
+ *  @brief  Render 2D axes.
  *  @param  object [in] pointer to object
  *  @param  camera [in] pointer to camera
  *  @param  light [in] pointer ot light
@@ -79,95 +84,111 @@ void Axis2D::exec( kvs::ObjectBase* object, kvs::Camera* camera, kvs::Light* lig
         //    +--------------------------+
         //
         const float dpr = camera->devicePixelRatio();
-        const int x0 = m_left_margin;
-        const int x1 = camera->windowWidth() - m_right_margin;
-        const int y0 = m_top_margin;
-        const int y1 = camera->windowHeight() - m_bottom_margin;
-//        const kvs::FontMetrics metrics = m_painter.fontMetrics();
+        const float x0 = m_left_margin;
+        const float x1 = camera->windowWidth() - m_right_margin;
+        const float y0 = m_top_margin;
+        const float y1 = camera->windowHeight() - m_bottom_margin;
+        const kvs::Vec4 rect( x0, x1, y0, y1 );
 
-        // Draw background.
-        if ( m_background_color.a() > 0.0f )
-        {
-            kvs::OpenGL::Begin( GL_QUADS );
-            kvs::OpenGL::Color( m_background_color );
-            kvs::OpenGL::Vertex( kvs::Vec2( x0, y0 ) * dpr );
-            kvs::OpenGL::Vertex( kvs::Vec2( x1, y0 ) * dpr );
-            kvs::OpenGL::Vertex( kvs::Vec2( x1, y1 ) * dpr );
-            kvs::OpenGL::Vertex( kvs::Vec2( x0, y1 ) * dpr );
-            kvs::OpenGL::End();
-        }
-
-        // Draw borderlines.
-        if ( m_borderline_color.a() > 0.0f )
-        {
-            const int d = int( m_borderline_width * 0.5 );
-            kvs::OpenGL::SetLineWidth( m_borderline_width * dpr );
-            kvs::OpenGL::Begin( GL_LINES );
-            kvs::OpenGL::Color( m_borderline_color );
-            kvs::OpenGL::Vertices( kvs::Vec2( x0 - d, y1 ) * dpr, kvs::Vec2( x1 + d, y1 ) * dpr ); // bottom
-            kvs::OpenGL::Vertices( kvs::Vec2( x0, y1 + d ) * dpr, kvs::Vec2( x0, y0 - d ) * dpr ); // left
-            kvs::OpenGL::Vertices( kvs::Vec2( x0 - d, y0 ) * dpr, kvs::Vec2( x1 + d, y0 ) * dpr ); // top
-            kvs::OpenGL::Vertices( kvs::Vec2( x1, y1 + d ) * dpr, kvs::Vec2( x1, y0 - d ) * dpr ); // right
-            kvs::OpenGL::End();
-        }
-
-        // Draw axes.
-        /*
-        const int d = int( m_axis_width * 0.5 );
-        kvs::OpenGL::SetLineWidth( m_axis_width * dpr );
-        kvs::OpenGL::Begin( GL_LINES );
-        kvs::OpenGL::Color( m_axis_color );
-        kvs::OpenGL::Vertices( kvs::Vec2( x0 - d, y1 ) * dpr, kvs::Vec2( x1 + d, y1 ) * dpr ); // X axis (bottom)
-        kvs::OpenGL::Vertices( kvs::Vec2( x0, y1 + d ) * dpr, kvs::Vec2( x0, y0 - d ) * dpr ); // Y axis (left)
-        kvs::OpenGL::Vertices( kvs::Vec2( x0 - d, y0 ) * dpr, kvs::Vec2( x1 + d, y0 ) * dpr ); // X axis (top)
-        kvs::OpenGL::Vertices( kvs::Vec2( x1, y1 + d ) * dpr, kvs::Vec2( x1, y0 - d ) * dpr ); // Y axis (right)
-        kvs::OpenGL::End();
-        */
+        // Draw background and borderlines.
+        this->draw_background( rect, dpr );
+        this->draw_borderline( rect, dpr );
 
         // Draw x-axis.
-        m_x_axis->setRect( kvs::Vec4i( x0, x1, y0, y1 ) );
+        if ( m_x_axis->label().empty() ) { m_x_axis->setLabel( table->label(0) ); }
+        m_x_axis->setRect( rect );
         m_x_axis->setRange( table->minValue(0), table->maxValue(0) );
-//        const int x_offset = 5;
-//        const int x_margin = m_x_axis->draw( m_painter, x_offset );
-        m_x_axis->setLabel( m_x_label.empty() ? table->label(0) : m_x_label );
-//        m_x_axis->setLabelOffset( 5 );
         m_x_axis->draw( m_painter );
 
-        m_y_axis->setRect( kvs::Vec4i( x0, x1, y0, y1 ) );
+        // Draw y-axis.
+        if ( m_y_axis->label().empty() ) { m_y_axis->setLabel( table->label(1) ); }
+        m_y_axis->setRect( rect );
         m_y_axis->setRange( table->minValue(1), table->maxValue(1) );
-//        const int y_offset = 5;
-//        const int y_margin = m_y_axis->draw( m_painter, y_offset );
-        m_y_axis->setLabel( m_y_label.empty() ? table->label(1) : m_y_label );
-//        m_y_axis->setLabelOffset( 5 );
         m_y_axis->draw( m_painter );
 
-        /*
-        // Draw x label.
-        const std::string x_label = m_x_label.empty() ? table->label(0) : m_x_label;
-        if ( x_label.size() > 0 )
-        {
-            const float px = ( x0 + x1 - metrics.width( x_label ) ) * 0.5f;
-            const float py = y1 + x_offset + x_margin + metrics.height();
-            m_painter.drawText( kvs::Vec2( px, py ), x_label );
-        }
-
-        // Draw y label.
-        const std::string y_label = m_y_label.empty() ? table->label(1) : m_y_label;
-        if ( y_label.size() > 0 )
-        {
-            const float px = x0 - y_offset - y_margin - y_offset;
-            const float py = ( y0 + y1 + metrics.width( y_label ) ) * 0.5f;
-            kvs::OpenGL::PushMatrix();
-            kvs::OpenGL::Translate( px * dpr, py * dpr, 0.0f );
-            kvs::OpenGL::Rotate( -90.0, 0, 0, 1 );
-            m_painter.drawText( kvs::Vec2( 0, 0 ), y_label );
-            kvs::OpenGL::PopMatrix();
-        }
-        */
+        // Draw title.
+        this->draw_title( rect );
     }
     m_painter.end();
 
     BaseClass::stopTimer();
+}
+
+/*===========================================================================*/
+/**
+ *  @brief  Draw the axis title
+ *  @param  rect [in] plot region
+ */
+/*===========================================================================*/
+void Axis2D::draw_title( const kvs::Vec4& rect )
+{
+    if ( m_title.size() > 0 )
+    {
+        const float x0 = rect[0];
+        const float x1 = rect[1];
+        const float y0 = rect[2];
+        const kvs::FontMetrics& metrics = m_painter.fontMetrics();
+        const kvs::Font font = m_painter.font();
+        {
+            m_painter.setFont( m_title_font );
+            const float px = ( x0 + x1 - metrics.width( m_title ) ) * 0.5f;
+            const float py = y0 - m_title_offset;
+            m_painter.drawText( kvs::Vec2( px, py ), m_title );
+        }
+        m_painter.setFont( font );
+    }
+}
+
+/*===========================================================================*/
+/**
+ *  @brief  Draw background of the plot region.
+ *  @param  rect [in] plot region
+ *  @param  dpr [in] device pixel ratio
+ */
+/*===========================================================================*/
+void Axis2D::draw_background( const kvs::Vec4& rect, const float dpr )
+{
+    if ( m_background_color.a() > 0.0f )
+    {
+        const float x0 = rect[0];
+        const float x1 = rect[1];
+        const float y0 = rect[2];
+        const float y1 = rect[3];
+        kvs::OpenGL::Begin( GL_QUADS );
+        kvs::OpenGL::Color( m_background_color );
+        kvs::OpenGL::Vertex( kvs::Vec2( x0, y0 ) * dpr );
+        kvs::OpenGL::Vertex( kvs::Vec2( x1, y0 ) * dpr );
+        kvs::OpenGL::Vertex( kvs::Vec2( x1, y1 ) * dpr );
+        kvs::OpenGL::Vertex( kvs::Vec2( x0, y1 ) * dpr );
+        kvs::OpenGL::End();
+    }
+}
+
+/*===========================================================================*/
+/**
+ *  @brief  Draw borderline of the plot region.
+ *  @param  rect [in] plot region
+ *  @param  dpr [in] device pixel ratio
+ */
+/*===========================================================================*/
+void Axis2D::draw_borderline( const kvs::Vec4& rect, const float dpr )
+{
+    if ( m_borderline_color.a() > 0.0f )
+    {
+        const float x0 = rect[0];
+        const float x1 = rect[1];
+        const float y0 = rect[2];
+        const float y1 = rect[3];
+        const int d = int( m_borderline_width * 0.5 );
+        kvs::OpenGL::SetLineWidth( m_borderline_width * dpr );
+        kvs::OpenGL::Begin( GL_LINES );
+        kvs::OpenGL::Color( m_borderline_color );
+        kvs::OpenGL::Vertices( kvs::Vec2( x0 - d, y1 ) * dpr, kvs::Vec2( x1 + d, y1 ) * dpr ); // bottom
+        kvs::OpenGL::Vertices( kvs::Vec2( x0, y1 + d ) * dpr, kvs::Vec2( x0, y0 - d ) * dpr ); // left
+        kvs::OpenGL::Vertices( kvs::Vec2( x0 - d, y0 ) * dpr, kvs::Vec2( x1 + d, y0 ) * dpr ); // top
+        kvs::OpenGL::Vertices( kvs::Vec2( x1, y1 + d ) * dpr, kvs::Vec2( x1, y0 - d ) * dpr ); // right
+        kvs::OpenGL::End();
+    }
 }
 
 } // end of namespace kvs