+++ Source/Core/Visualization/Renderer/ValueAxis.cpp
@@ -69,27 +69,36 @@ void ValueAxis::draw( kvs::Painter& painter )
 /*===========================================================================*/
 int ValueAxis::drawTickLabelsOnTop( kvs::Painter& painter )
 {
+    if ( !m_tick_label_visible ) { return 0; } // invisible
+
     const float x0 = m_rect[0];
     const float x1 = m_rect[1];
     const float y0 = m_rect[2];
-    const kvs::FontMetrics& metrics = painter.fontMetrics();
 
     // Format
     const int precision = ( m_precision == 0 ) ? 6 : m_precision;
     const bool fixed = ( m_precision == 0 ) ? false : true;
     const bool scientific = m_scientific;
 
-    const float stride = ( x1 - x0 ) / ( m_nticks - 1 );
-    for ( size_t i = 0; i < m_nticks; ++i )
+    int margin = 0;
+    const kvs::Font font = painter.font();
+    painter.setFont( m_tick_label_font );
     {
-        const double value = m_min + i * ( m_max - m_min ) / ( m_nticks - 1 );
-        const auto v = kvs::String::ToString( value, precision, fixed, scientific );
-        const int x = x0 + stride * i - metrics.width( v ) / 2;
-        const int y = y0 - metrics.height() - m_label_offset;
-        painter.drawText( kvs::Vec2( x, y ), v );
+        const kvs::FontMetrics& metrics = painter.fontMetrics();
+        const float dx = ( x1 - x0 ) / ( m_nticks - 1 );
+        for ( size_t i = 0; i < m_nticks; ++i )
+        {
+            const double value = m_min + i * ( m_max - m_min ) / ( m_nticks - 1 );
+            const auto v = kvs::String::ToString( value, precision, fixed, scientific );
+            const int x = x0 + dx * i - metrics.width( v ) / 2;
+            const int y = y0 - metrics.height() - m_label_offset;
+            painter.drawText( kvs::Vec2( x, y ), v );
+        }
+        margin = metrics.height();
     }
+    painter.setFont( font );
 
-    return metrics.height();
+    return margin;
 }
 
 /*===========================================================================*/
@@ -101,27 +110,36 @@ int ValueAxis::drawTickLabelsOnTop( kvs::Painter& painter )
 /*===========================================================================*/
 int ValueAxis::drawTickLabelsOnBottom( kvs::Painter& painter )
 {
+    if ( !m_tick_label_visible ) { return 0; } // invisible
+
     const float x0 = m_rect[0];
     const float x1 = m_rect[1];
     const float y1 = m_rect[3];
-    const kvs::FontMetrics& metrics = painter.fontMetrics();
 
     // Format
     const int precision = ( m_precision == 0 ) ? 6 : m_precision;
     const bool fixed = ( m_precision == 0 ) ? false : true;
     const bool scientific = m_scientific;
 
-    const float stride = ( x1 - x0 ) / ( m_nticks - 1 );
-    for ( size_t i = 0; i < m_nticks; ++i )
+    int margin = 0;
+    const kvs::Font font = painter.font();
+    painter.setFont( m_tick_label_font );
     {
-        const double value = m_min + i * ( m_max - m_min ) / ( m_nticks - 1 );
-        const auto v = kvs::String::ToString( value, precision, fixed, scientific );
-        const int x = x0 + stride * i - metrics.width( v ) / 2;
-        const int y = y1 + metrics.height() + m_label_offset;
-        painter.drawText( kvs::Vec2( x, y ), v );
+        const kvs::FontMetrics& metrics = painter.fontMetrics();
+        const float dx = ( x1 - x0 ) / ( m_nticks - 1 );
+        for ( size_t i = 0; i < m_nticks; ++i )
+        {
+            const double value = m_min + i * ( m_max - m_min ) / ( m_nticks - 1 );
+            const auto v = kvs::String::ToString( value, precision, fixed, scientific );
+            const int x = x0 + dx * i - metrics.width( v ) / 2;
+            const int y = y1 + metrics.height() + m_label_offset;
+            painter.drawText( kvs::Vec2( x, y ), v );
+        }
+        margin = metrics.height();
     }
+    painter.setFont( font );
 
-    return metrics.height();
+    return margin;
 }
 
 /*===========================================================================*/
@@ -133,10 +151,11 @@ int ValueAxis::drawTickLabelsOnBottom( kvs::Painter& painter )
 /*===========================================================================*/
 int ValueAxis::drawTickLabelsOnLeft( kvs::Painter& painter )
 {
+    if ( !m_tick_label_visible ) { return 0; } // invisible
+
     const float x0 = m_rect[0];
     const float y0 = m_rect[2];
     const float y1 = m_rect[3];
-    const kvs::FontMetrics& metrics = painter.fontMetrics();
     const int offset = 3; // additional offset between the axis and tick label
 
     // Format
@@ -145,18 +164,25 @@ int ValueAxis::drawTickLabelsOnLeft( kvs::Painter& painter )
     const bool scientific = m_scientific;
 
     int margin = 0;
-    const float stride = ( y1 - y0 ) / ( m_nticks - 1 );
-    for ( size_t i = 0; i < m_nticks; ++i )
+    const kvs::Font font = painter.font();
+    painter.setFont( m_tick_label_font );
     {
-        const double value = m_min + i * ( m_max - m_min ) / ( m_nticks - 1 );
-        const auto v = kvs::String::ToString( value, precision, fixed, scientific );
-        const int x = x0 - metrics.width( v ) - m_label_offset - offset;
-        const int y = y1 - stride * i + metrics.height() / 2;
-        painter.drawText( kvs::Vec2( x, y ), v );
-        margin = kvs::Math::Max( margin, metrics.width( v ) );
+        const kvs::FontMetrics& metrics = painter.fontMetrics();
+        const float dy = ( y1 - y0 ) / ( m_nticks - 1 );
+        for ( size_t i = 0; i < m_nticks; ++i )
+        {
+            const double value = m_min + i * ( m_max - m_min ) / ( m_nticks - 1 );
+            const auto v = kvs::String::ToString( value, precision, fixed, scientific );
+            const int x = x0 - metrics.width( v ) - m_label_offset - offset;
+            const int y = y1 - dy * i + metrics.height() / 2;
+            painter.drawText( kvs::Vec2( x, y ), v );
+            margin = kvs::Math::Max( margin, metrics.width( v ) );
+        }
+        margin += offset;
     }
+    painter.setFont( font );
 
-    return margin + offset;
+    return margin;
 }
 
 /*===========================================================================*/
@@ -168,10 +194,11 @@ int ValueAxis::drawTickLabelsOnLeft( kvs::Painter& painter )
 /*===========================================================================*/
 int ValueAxis::drawTickLabelsOnRight( kvs::Painter& painter )
 {
+    if ( !m_tick_label_visible ) { return 0; } // invisible
+
     const float x1 = m_rect[1];
     const float y0 = m_rect[2];
     const float y1 = m_rect[3];
-    const kvs::FontMetrics& metrics = painter.fontMetrics();
     const int offset = 3; // additional offset between the axis and tick label
 
     // Format
@@ -180,18 +207,25 @@ int ValueAxis::drawTickLabelsOnRight( kvs::Painter& painter )
     const bool scientific = m_scientific;
 
     int margin = 0;
-    const float stride = ( y1 - y0 ) / ( m_nticks - 1 );
-    for ( size_t i = 0; i < m_nticks; ++i )
+    const kvs::Font font = painter.font();
+    painter.setFont( m_tick_label_font );
     {
-        const double value = m_min + i * ( m_max - m_min ) / ( m_nticks - 1 );
-        const auto v = kvs::String::ToString( value, precision, fixed, scientific );
-        const int x = x1 + m_label_offset + offset;
-        const int y = y1 - stride * i + metrics.height() / 2;
-        painter.drawText( kvs::Vec2( x, y ), v );
-        margin = kvs::Math::Max( margin, metrics.width( v ) );
+        const kvs::FontMetrics& metrics = painter.fontMetrics();
+        const float dy = ( y1 - y0 ) / ( m_nticks - 1 );
+        for ( size_t i = 0; i < m_nticks; ++i )
+        {
+            const double value = m_min + i * ( m_max - m_min ) / ( m_nticks - 1 );
+            const auto v = kvs::String::ToString( value, precision, fixed, scientific );
+            const int x = x1 + m_label_offset + offset;
+            const int y = y1 - dy * i + metrics.height() / 2;
+            painter.drawText( kvs::Vec2( x, y ), v );
+            margin = kvs::Math::Max( margin, metrics.width( v ) );
+        }
+        margin += offset;
     }
+    painter.setFont( font );
 
-    return margin + offset;
+    return margin;
 }
 
 /*===========================================================================*/
@@ -202,6 +236,8 @@ int ValueAxis::drawTickLabelsOnRight( kvs::Painter& painter )
 /*===========================================================================*/
 void ValueAxis::drawTickMarksOnTop( kvs::Painter& painter )
 {
+    if ( !m_tick_mark_visible ) { return; } // invisible
+
     const float x0 = m_rect[0];
     const float x1 = m_rect[1];
     const float y0 = m_rect[2];
@@ -209,7 +245,7 @@ void ValueAxis::drawTickMarksOnTop( kvs::Painter& painter )
 
     float x = x0 * dpr;
     const float y = y0 * dpr;
-    const float dy = m_tick_mark_length * dpr;
+    const float dy = m_tick_direction * m_tick_mark_length * dpr;
     const float dx = ( ( x1 - x0 ) / ( m_nticks - 1 ) ) * dpr;
     for ( size_t i = 0; i < m_nticks; ++i, x += dx )
     {
@@ -229,6 +265,8 @@ void ValueAxis::drawTickMarksOnTop( kvs::Painter& painter )
 /*===========================================================================*/
 void ValueAxis::drawTickMarksOnBottom( kvs::Painter& painter )
 {
+    if ( !m_tick_mark_visible ) { return; } // invisible
+
     const float x0 = m_rect[0];
     const float x1 = m_rect[1];
     const float y1 = m_rect[3];
@@ -236,7 +274,7 @@ void ValueAxis::drawTickMarksOnBottom( kvs::Painter& painter )
 
     float x = x0 * dpr;
     const float y = y1 * dpr;
-    const float dy = m_tick_mark_length * dpr;
+    const float dy = m_tick_direction * m_tick_mark_length * dpr;
     const float dx = ( ( x1 - x0 ) / ( m_nticks - 1 ) ) * dpr;
     for ( size_t i = 0; i < m_nticks; ++i, x += dx )
     {
@@ -256,6 +294,8 @@ void ValueAxis::drawTickMarksOnBottom( kvs::Painter& painter )
 /*===========================================================================*/
 void ValueAxis::drawTickMarksOnLeft( kvs::Painter& painter )
 {
+    if ( !m_tick_mark_visible ) { return; } // invisible
+
     const float x0 = m_rect[0];
     const float y0 = m_rect[2];
     const float y1 = m_rect[3];
@@ -263,7 +303,7 @@ void ValueAxis::drawTickMarksOnLeft( kvs::Painter& painter )
 
     float y = y1 * dpr;
     const float x = x0 * dpr;
-    const float dx = m_tick_mark_length * dpr;
+    const float dx = m_tick_direction * m_tick_mark_length * dpr;
     const float dy = ( ( y1 - y0 ) / ( m_nticks - 1 ) ) * dpr;
     for ( size_t i = 0; i < m_nticks; ++i, y -= dy )
     {
@@ -283,6 +323,8 @@ void ValueAxis::drawTickMarksOnLeft( kvs::Painter& painter )
 /*===========================================================================*/
 void ValueAxis::drawTickMarksOnRight( kvs::Painter& painter )
 {
+    if ( !m_tick_mark_visible ) { return; } // invisible
+
     const float x1 = m_rect[1];
     const float y0 = m_rect[2];
     const float y1 = m_rect[3];
@@ -290,7 +332,7 @@ void ValueAxis::drawTickMarksOnRight( kvs::Painter& painter )
 
     float y = y1 * dpr;
     const float x = x1 * dpr;
-    const float dx = m_tick_mark_length * dpr;
+    const float dx = m_tick_direction * m_tick_mark_length * dpr;
     const float dy = ( ( y1 - y0 ) / ( m_nticks - 1 ) ) * dpr;
     for ( size_t i = 0; i < m_nticks; ++i, y -= dy )
     {
@@ -310,6 +352,8 @@ void ValueAxis::drawTickMarksOnRight( kvs::Painter& painter )
 /*===========================================================================*/
 void ValueAxis::drawAxisOnTop( kvs::Painter& painter )
 {
+    if ( !m_visible ) { return; } // invisible
+
     const float x0 = m_rect[0];
     const float x1 = m_rect[1];
     const float y0 = m_rect[2];
@@ -330,6 +374,8 @@ void ValueAxis::drawAxisOnTop( kvs::Painter& painter )
 /*===========================================================================*/
 void ValueAxis::drawAxisOnBottom( kvs::Painter& painter )
 {
+    if ( !m_visible ) { return; } // invisible
+
     const float x0 = m_rect[0];
     const float x1 = m_rect[1];
     const float y1 = m_rect[3];
@@ -350,6 +396,8 @@ void ValueAxis::drawAxisOnBottom( kvs::Painter& painter )
 /*===========================================================================*/
 void ValueAxis::drawAxisOnLeft( kvs::Painter& painter )
 {
+    if ( !m_visible ) { return; } // invisible
+
     const float x0 = m_rect[0];
     const float y0 = m_rect[2];
     const float y1 = m_rect[3];
@@ -370,6 +418,8 @@ void ValueAxis::drawAxisOnLeft( kvs::Painter& painter )
 /*===========================================================================*/
 void ValueAxis::drawAxisOnRight( kvs::Painter& painter )
 {
+    if ( !m_visible ) { return; } // invisible
+
     const float x1 = m_rect[1];
     const float y0 = m_rect[2];
     const float y1 = m_rect[3];
@@ -391,16 +441,22 @@ void ValueAxis::drawAxisOnRight( kvs::Painter& painter )
 /*===========================================================================*/
 void ValueAxis::drawLabelOnTop( kvs::Painter& painter, const int margin )
 {
-    const int x0 = m_rect[0];
-    const int x1 = m_rect[1];
-    const int y0 = m_rect[2];
-    const kvs::FontMetrics& metrics = painter.fontMetrics();
-    if ( m_label.size() > 0 )
+    if ( !m_label_visible ) { return; } // invisible
+    if ( m_label.size() == 0 ) { return; } // empty
+
+    const float x0 = m_rect[0];
+    const float x1 = m_rect[1];
+    const float y0 = m_rect[2];
+
+    const kvs::Font font = painter.font();
+    painter.setFont( m_label_font );
     {
+        const kvs::FontMetrics& metrics = painter.fontMetrics();
         const float px = ( x0 + x1 - metrics.width( m_label ) ) * 0.5f;
         const float py = y0 - m_label_offset - margin - metrics.height();
         painter.drawText( kvs::Vec2( px, py ), m_label );
-    }
+        }
+    painter.setFont( font );
 }
 
 /*===========================================================================*/
@@ -412,16 +468,22 @@ void ValueAxis::drawLabelOnTop( kvs::Painter& painter, const int margin )
 /*===========================================================================*/
 void ValueAxis::drawLabelOnBottom( kvs::Painter& painter, const int margin )
 {
-    const int x0 = m_rect[0];
-    const int x1 = m_rect[1];
-    const int y1 = m_rect[3];
-    const kvs::FontMetrics& metrics = painter.fontMetrics();
-    if ( m_label.size() > 0 )
+    if ( !m_label_visible ) { return; } // invisible
+    if ( m_label.size() == 0 ) { return; } // empty
+
+    const float x0 = m_rect[0];
+    const float x1 = m_rect[1];
+    const float y1 = m_rect[3];
+
+    const kvs::Font font = painter.font();
+    painter.setFont( m_label_font );
     {
+        const kvs::FontMetrics& metrics = painter.fontMetrics();
         const float px = ( x0 + x1 - metrics.width( m_label ) ) * 0.5f;
         const float py = y1 + m_label_offset + margin + metrics.height();
         painter.drawText( kvs::Vec2( px, py ), m_label );
     }
+    painter.setFont( font );
 }
 
 /*===========================================================================*/
@@ -433,13 +495,18 @@ void ValueAxis::drawLabelOnBottom( kvs::Painter& painter, const int margin )
 /*===========================================================================*/
 void ValueAxis::drawLabelOnLeft( kvs::Painter& painter,  const int margin )
 {
-    const int x0 = m_rect[0];
-    const int y0 = m_rect[2];
-    const int y1 = m_rect[3];
+    if ( !m_label_visible ) { return; } // invisible
+    if ( m_label.size() == 0 ) { return; } // empty
+
+    const float x0 = m_rect[0];
+    const float y0 = m_rect[2];
+    const float y1 = m_rect[3];
     const float dpr = painter.devicePixelRatio();
-    const kvs::FontMetrics& metrics = painter.fontMetrics();
-    if ( m_label.size() > 0 )
+
+    const kvs::Font font = painter.font();
+    painter.setFont( m_label_font );
     {
+        const kvs::FontMetrics& metrics = painter.fontMetrics();
         const float px = x0 - m_label_offset - margin - m_label_offset;
         const float py = ( y0 + y1 + metrics.width( m_label ) ) * 0.5f;
         kvs::OpenGL::PushMatrix();
@@ -448,6 +515,7 @@ void ValueAxis::drawLabelOnLeft( kvs::Painter& painter,  const int margin )
         painter.drawText( kvs::Vec2( 0, 0 ), m_label );
         kvs::OpenGL::PopMatrix();
     }
+    painter.setFont( font );
 }
 
 /*===========================================================================*/
@@ -459,13 +527,18 @@ void ValueAxis::drawLabelOnLeft( kvs::Painter& painter,  const int margin )
 /*===========================================================================*/
 void ValueAxis::drawLabelOnRight( kvs::Painter& painter, const int margin )
 {
-    const int x1 = m_rect[1];
-    const int y0 = m_rect[2];
-    const int y1 = m_rect[3];
+    if ( !m_label_visible ) { return; } // invisible
+    if ( m_label.size() == 0 ) { return; } // empty
+
+    const float x1 = m_rect[1];
+    const float y0 = m_rect[2];
+    const float y1 = m_rect[3];
     const float dpr = painter.devicePixelRatio();
-    const kvs::FontMetrics& metrics = painter.fontMetrics();
-    if ( m_label.size() > 0 )
+
+    const kvs::Font font = painter.font();
+    painter.setFont( m_label_font );
     {
+        const kvs::FontMetrics& metrics = painter.fontMetrics();
         const float px = x1 + m_label_offset + margin + m_label_offset;
         const float py = ( y0 + y1 + metrics.width( m_label ) ) * 0.5f;
         kvs::OpenGL::PushMatrix();
@@ -474,6 +547,7 @@ void ValueAxis::drawLabelOnRight( kvs::Painter& painter, const int margin )
         painter.drawText( kvs::Vec2( 0, 0 ), m_label );
         kvs::OpenGL::PopMatrix();
     }
+    painter.setFont( font );
 }
 
 } // end of namespace kvs